requirements.txt
attrs==24.2.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.32.3
requests-cache==1.2.1
six==1.16.0
url-normalize==1.4.3
urllib3==2.2.3
